src/core/services/DesignService.ts
import { S3Client } from '@aws-sdk/client-s3';
import { env, constants } from '../../config/config';
import { Design, DesignContent, DesignStatus } from '../models/types';
import { Logger } from './Logger';
import { CloudflareR2Service } from '../../services/CloudflareR2Service';

/**
 * Service for managing design storage and retrieval
 */
export class DesignService {
  private logger: Logger;
  private storage: CloudflareR2Service;

  constructor() {
    this.logger = new Logger('DesignService');
    this.storage = new CloudflareR2Service();
  }

  /**
   * Create a new design
   */
  public async createDesign(userId: string, title: string, content: DesignContent): Promise<Design> {
    try {
      const design: Design = {
        id: crypto.randomUUID(),
        userId,
        title,
        description: '',
        content,
        version: 1,
        status: DesignStatus.DRAFT,
        createdAt: new Date(),
        updatedAt: new Date(),
      };

      // Store design content in R2
      await this.storage.uploadDesign(design.id, JSON.stringify(content));

      // TODO: Store design metadata in database
      this.logger.info(`Created design ${design.id} for user ${userId}`);

      return design;
    } catch (error) {
      this.logger.error('Error creating design:', error);
      throw new Error('Failed to create design');
    }
  }

  /**
   * Retrieve a design by ID
   */
  public async getDesign(designId: string): Promise<Design> {
    try {
      // TODO: Retrieve design metadata from database
      const contentJson = await this.storage.getDesign(designId);
      const content = JSON.parse(contentJson) as DesignContent;

      // Temporary mock implementation
      const design: Design = {
        id: designId,
        userId: 'mock-user',
        title: 'Mock Design',
        description: '',
        content,
        version: 1,
        status: DesignStatus.DRAFT,
        createdAt: new Date(),
        updatedAt: new Date(),
      };

      return design;
    } catch (error) {
      this.logger.error(`Error retrieving design ${designId}:`, error);
      throw new Error('Failed to retrieve design');
    }
  }

  /**
   * Update an existing design
   */
  public async updateDesign(designId: string, updates: Partial<Design>): Promise<Design> {
    try {
      const currentDesign = await this.getDesign(designId);
      
      const updatedDesign: Design = {
        ...currentDesign,
        ...updates,
        version: currentDesign.version + 1,
        updatedAt: new Date(),
      };

      if (updates.content) {
        await this.storage.uploadDesign(designId, JSON.stringify(updatedDesign.content));
      }

      // TODO: Update design metadata in database
      this.logger.info(`Updated design ${designId}`);

      return updatedDesign;
    } catch (error) {
      this.logger.error(`Error updating design ${designId}:`, error);
      throw new Error('Failed to update design');
    }
  }

  /**
   * Delete a design
   */
  public async deleteDesign(designId: string): Promise<void> {
    try {
      await this.storage.deleteDesign(designId);
      // TODO: Delete design metadata from database
      this.logger.info(`Deleted design ${designId}`);
    } catch (error) {
      this.logger.error(`Error deleting design ${designId}:`, error);
      throw new Error('Failed to delete design');
    }
  }

  /**
   * List designs for a user
   */
  public async listDesigns(userId: string, status?: DesignStatus): Promise<Design[]> {
    try {
      // TODO: Implement database query for user's designs
      this.logger.info(`Listing designs for user ${userId}`);
      return []; // Temporary empty response
    } catch (error) {
      this.logger.error(`Error listing designs for user ${userId}:`, error);
      throw new Error('Failed to list designs');
    }
  }

  /**
   * Export design in specified format
   */
  public async exportDesign(designId: string, format: string): Promise<Buffer> {
    try {
      const design = await this.getDesign(designId);
      // TODO: Implement export logic for different formats
      this.logger.info(`Exporting design ${designId} in ${format} format`);
      return Buffer.from(''); // Temporary empty response
    } catch (error) {
      this.logger.error(`Error exporting design ${designId}:`, error);
      throw new Error('Failed to export design');
    }
  }
}

// Export singleton instance
export const designService = new DesignService(); 